pyproject.toml
[tool.poetry]
name = "grpc-minio"
version = "0.1.0"
description = ""
authors = ["yang-lin <casongg@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.65.1"
grpcio-tools = "^1.65.1"
minio = "^7.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
